requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
transformers>=4.35.0
torch>=2.0.0
supabase>=2.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0